requirements.txt
deep_translator==1.11.1
diffusers==0.17.1
keybert==0.7.0
keyphrase_vectorizers==0.0.11
matplotlib==3.7.1
nltk==3.8.1
Pillow==9.5.0
Requests==2.31.0
spacy==3.5.3
torch==2.0.0+cu118
transformers==4.30.2
wordcloud==1.9.2
